requirements.txt
pyyaml
urllib3
requests

